The Four Foxes new version
Revised rules for fear and damage, they are simpler and more narrative.
Improvements on the Diceless sections for Players and GM.
Started sections for Case creation
First case description.

The two sections on case creation are highly experimental. I'd like to do something similar to the Transmissions from Technoir, but simpler.